% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertEO.R
\name{insertEO}
\alias{insertEO}
\title{Auxiliary function that put together even and odd functions of the SVD in an only one output list.}
\usage{
insertEO(Psi_odd, Psi_even)
}
\arguments{
\item{Psi_odd}{odd singular functions Psi}

\item{Psi_even}{even singular functions Psi}
}
\value{
Psi
}
\description{
Auxiliary function that put together even and odd functions of the SVD in an only one output list.
}
\examples{
library(orthopolynom)
library(polynom)
library(tmvtnorm)
library(ks)
library(sfsmisc)
library(snowfall)
library(fourierin)
library(rdetools)
library(statmod)
library(RCEIM)
library(robustbase)
library(VGAM)
library(RandomCoefficients)
#### Number of Psis
L =15
L1 = L+1
N2 = max(L,3)
twoN = 2*N2
#### Bandwidth 1
c1 = 1
K1 = max(twoN+2,30)
K = K1
### get the beta's (for the computation of Psi)
mu1<- MU(N2,c1,K1, L1)
output <- insertEO(mu1[[2]], mu1[[1]])

}
