% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDiag.R
\name{myDiag}
\alias{myDiag}
\title{Auxiliary function to form matrices equal to  x everywhere except from the upper/lower k diagonal, which values are vec}
\usage{
myDiag(x, vec, k)
}
\arguments{
\item{x}{initial matriw}

\item{vec}{vector with specified new values for the upper/lower k diagonal elements of x}

\item{k}{index of the diagonal}
}
\value{
x
}
\description{
Auxiliary function to form matrices equal to  x everywhere except from the upper/lower k diagonal, which values are vec
}
\examples{
K=3
u <- sqrt(1/(4-1/seq(1,(K-1))^2))
n = length(u)+1
trans = myDiag(matrix(0,n, n),u,1) + myDiag(matrix(0,n, n),u,-1)

}
